requirements.txt
absl-py==0.11.0
alabaster==0.7.12
appdirs==1.4.4
argon2-cffi==20.1.0
astroid==2.4.2
async-generator==1.10
attrs==20.3.0
awscli==1.19.6
Babel==2.9.0
backcall==0.2.0
black==20.8b1
bleach==3.2.1
botocore==1.20.6
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.3
coverage==5.3.1
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
grpcio==1.34.1
idna==2.10
imageio==2.9.0
imagesize==1.2.0
iniconfig==1.1.1
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.7.0
jedi==0.18.0
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
networkx==2.5
notebook==6.1.6
numpy==1.20.1
oauthlib==3.1.0
packaging==20.8
pandas==1.2.0
pandocfilters==1.4.3
parso==0.8.1
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
pluggy==0.13.1
prometheus-client==0.9.0
prompt-toolkit==3.0.10
protobuf==3.14.0
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.20
Pygments==2.7.3
pylint==2.6.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.1
pytest-cov==2.10.1
python-dateutil==2.8.1
pytorch-msssim==0.2.0
pytorch-gradcam==0.2.1
pytz==2020.5
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.5
s3transfer==0.3.4
scikit-image==0.18.1
scipy==1.6.0
Send2Trash==1.5.0
six==1.15.0
snowballstemmer==2.0.0
Sphinx==3.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
tensorboard==2.4.1
tensorboard-plugin-wit==1.7.0
terminado==0.9.2
testpath==0.4.4
tifffile==2021.2.1
toml==0.10.2
tornado==6.1
tqdm==4.56.0
traitlets==5.0.5
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
